update agent; tool calling needs tuning; login components
